Add files via upload
Hey guys. Just an update this is the version I have been working on for our project. It should be done tomorrow and will have more functionality. First off the first window that pops up in the top left corner of the window is a message Pane from JOptionpane. It disappears within 10 seconds and prompts another box with user interactivity and will soon take input and spit out an answer. Take a look! (as of right now it will not give an answer so just wait I should have that up and running by tomorrow!)

**There is a main method file and a method body file in this upload**